stock/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alpaca-py==0.39.1
alpaca-trade-api==3.2.0
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
deprecation==2.1.0
frozenlist==1.5.0
idna==3.10
msgpack==1.0.3
multidict==6.3.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
propcache==0.3.1
pydantic==2.11.2
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.1
requests==2.32.3
six==1.17.0
sseclient-py==1.8.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==1.26.20
websocket-client==1.8.0
websockets==10.4
yarl==1.18.3
